+++ src/main/java/gdv/xport/satz/VertragsspezifischerTeil.java
@@ -116,7 +116,7 @@ public class VertragsspezifischerTeil extends Datensatz {
      */
     @Override
     public void setSparte(int x) {
-        if (this.getSatzart() == x) {
+        if (this.getSatzart().toInt() == x) {
             log.debug("nothing to do here - old Sparte = new Sparte (" + x + ")");
         }
         super.setSparte(x);
