+++ src/test/java/gdv/xport/satz/AllgemeinerVertragsteilTest.java
@@ -35,7 +35,7 @@ public class AllgemeinerVertragsteilTest extends AbstractSatzTest {
      */
     @Test
     public void testAllgemeinerVertragsteil() {
-        assertEquals(200, vertragsteil.getSatzart());
+        assertEquals(200, vertragsteil.getSatzart().toInt());
     }
 
 }
